+++ src/sqlic/mod.rs
@@ -1,2 +1,3 @@
 
 pub mod sql_utils;
+pub mod sql_crud;